requirements.txt
matrix-nio[e2e]>=0.20.1
thefuzz>=0.19.0
mistune==3.0.1
PyYAML>=6.0
jsonpickle>=2.1.0
Pillow==9.5.0
blurhash-python==1.2.0
requests==2.32.0
uvloop==0.21.0
levenshtein==0.26.1
